# openapi.yaml
...
openapi: 3.1.0
info:
  title: LSI-Maintenance — API Contrats
  version: 1.0.0
  description: "API de lecture des contrats clients pour les applications de la suite LSI-Maintenance. Authentification par clé d’API (`Authorization: Bearer ctr_…`), scopes fins, débit limité par client. Pagination par curseur (`nextCursor`), `ETag` / `If-None-Match`, erreurs RFC 9457. Guide : docs/contrats/07-api.md."
servers:
  - url: https://contrats.lsi-maintenance.fr
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: ctr_<prefix>_<secret>
      description: "Clé créée par un administrateur (Paramètres → API). Scopes : contracts:read, contracts:dates:read, pricing:read, pricing:quote, webhooks:manage, proposals:read, proposals:pricing:read."
  schemas:
    Contract:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        reference:
          type: string
        title:
          type: string
        type:
          type: string
          enum:
            - MAIN
            - AMENDMENT
        status:
          type: string
          enum:
            - DRAFT
            - IN_REVIEW
            - CHANGES_REQUESTED
            - APPROVED
            - SENT_TO_CLIENT
            - IN_NEGOTIATION
            - ACCEPTED
            - PENDING_SIGNATURE
            - PARTIALLY_SIGNED
            - SIGNATURE_EXPIRED
            - SIGNED
            - ACTIVE
            - RENEWAL_DUE
            - TERMINATION_PENDING
            - EXPIRED
            - TERMINATED
            - RENEWED
            - CANCELLED
            - DECLINED
            - IMPORTED_PENDING_VALIDATION
        origin:
          type: string
          enum:
            - NATIVE
            - IMPORTED
            - AI
            - PROPOSAL
        category:
          type: string
        customer:
          type: object
          properties:
            id:
              type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            name:
              type: string
            externalRef:
              anyOf:
                - type: string
                - type: "null"
            siren:
              anyOf:
                - type: string
                - type: "null"
          required:
            - id
            - name
            - externalRef
            - siren
          additionalProperties: false
        parentContractId:
          anyOf:
            - type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            - type: "null"
        currency:
          type: string
        startDate:
          anyOf:
            - type: string
              format: date
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
              description: Date calendaire AAAA-MM-JJ
            - type: "null"
        endDate:
          anyOf:
            - type: string
              format: date
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
              description: Date calendaire AAAA-MM-JJ
            - type: "null"
        renewalMode:
          type: string
          enum:
            - NONE
            - TACIT
            - EXPRESS
        signatureMode:
          anyOf:
            - type: string
              enum:
                - PDF
                - TEMPLATE
            - type: "null"
          description: Mode de la dernière demande de signature ; null si aucune.
        signedAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
            - type: "null"
        activatedAt:
          anyOf:
            - type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
            - type: "null"
        updatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
      required:
        - id
        - reference
        - title
        - type
        - status
        - origin
        - category
        - customer
        - parentContractId
        - currency
        - startDate
        - endDate
        - renewalMode
        - signatureMode
        - signedAt
        - activatedAt
        - updatedAt
      additionalProperties: false
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
      required:
        - type
        - title
        - status
        - detail
        - instance
        - code
      additionalProperties: {}
      description: Erreur RFC 9457 (`application/problem+json`).
...
paths:
...
  /api/v1/contracts/{id}:
    get:
      operationId: getContract
      summary: "Détail d’un contrat : statut, origine, mode de signature"
      tags:
        - Contrats
      security:
        - apiKey:
            - contracts:read
      x-required-scope: contracts:read
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du contrat
          schema:
            type: string
            format: uuid
        - name: If-None-Match
          in: header
          required: false
          description: ETag d’une réponse précédente → 304 si inchangée.
          schema:
            type: string
      responses:
        "200":
          description: Succès
          headers:
            ETag:
              schema:
                type: string
            RateLimit-Remaining:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "304":
          description: Non modifié (If-None-Match)
        "400":
          description: Requête invalide
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Clé d’API absente ou invalide
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Scope insuffisant ou API désactivée
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Ressource introuvable (ou hors du tenant)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Débit dépassé (`Retry-After`)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...